requirements.txt
customtkinter==5.2.2
pillow==10.4.0
sentence-transformers==3.1.1
chromadb==0.5.23
ollama==0.4.4
pypdf==5.0.1
python-docx==1.1.2
pyinstaller==6.10.0
httpx==0.27.2
langchain==0.3.7
langchain-community==0.3.7
python-telegram-bot==21.7
apscheduler==3.10.4
SpeechRecognition==3.10.0
PyAudio==0.2.14
